services/graphhopper-service/graphhopper-service/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
requests==2.31.0
redis==5.0.1
psycopg2-binary==2.9.7
schedule==1.2.0
polyline==2.0.0
geopy==2.3.0
